requirements.txt
# Core dependencies
rich>=13.0.0
python-dotenv>=1.0.0
requests>=2.30.0
pydantic>=2.0.0

# Optional: for extended functionality
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0